+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/application/v6/SiteUpdateApplicationV6Service.kt
@@ -30,8 +30,8 @@ import org.eclipse.tractusx.bpdm.pool.mapper.poolv6.outbound.SiteParseErrorMappe
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv6.outbound.SiteResponseMapperV6
 import org.eclipse.tractusx.bpdm.pool.model.ParseResult
 import org.eclipse.tractusx.bpdm.pool.model.parseAndExecute
-import org.eclipse.tractusx.bpdm.pool.service.operation.SitePayloadUpdateService
-import org.eclipse.tractusx.bpdm.pool.service.parser.SiteUpdateParser
+import org.eclipse.tractusx.bpdm.pool.service.operation.site.SitePayloadUpdateService
+import org.eclipse.tractusx.bpdm.pool.service.parser.site.SiteUpdateParser
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional