+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/application/v7/AddressSearchApplicationV7Service.kt
@@ -29,9 +29,9 @@ import org.eclipse.tractusx.bpdm.pool.api.model.LogisticAddressVerboseDto
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv7.inbound.AddressSearchRequestMapper
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv7.outbound.AddressResponseMapper
 import org.eclipse.tractusx.bpdm.pool.model.ParseResult
-import org.eclipse.tractusx.bpdm.pool.service.operation.AddressSearchService
-import org.eclipse.tractusx.bpdm.pool.service.parser.AddressSearchParser
-import org.eclipse.tractusx.bpdm.pool.service.parser.LegalEntityAddressSearchParser
+import org.eclipse.tractusx.bpdm.pool.service.operation.address.AddressSearchService
+import org.eclipse.tractusx.bpdm.pool.service.parser.address.AddressSearchParser
+import org.eclipse.tractusx.bpdm.pool.service.parser.legalentity.LegalEntityAddressSearchParser
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.eclipse.tractusx.bpdm.pool.api.model.request.AddressSearchRequest as AddressSearchRequestDto