+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/application/v7/ChangelogSearchApplicationV7Service.kt
@@ -28,7 +28,7 @@ import org.eclipse.tractusx.bpdm.pool.mapper.poolv7.inbound.ChangelogSearchReque
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv7.outbound.ChangelogResponseMapper
 import org.eclipse.tractusx.bpdm.pool.mapper.shared.outbound.ChangelogParseErrorMapper
 import org.eclipse.tractusx.bpdm.pool.model.ParseResult
-import org.eclipse.tractusx.bpdm.pool.service.operation.ChangelogSearchService
+import org.eclipse.tractusx.bpdm.pool.service.operation.changelog.ChangelogSearchService
 import org.eclipse.tractusx.bpdm.pool.service.parser.ChangelogSearchParser
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional