+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/application/v7/DataSpaceParticipantUpdateApplicationV7Service.kt
@@ -23,8 +23,8 @@ import mu.KotlinLogging
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv7.inbound.DataSpaceParticipantUpdateRequestMapper
 import org.eclipse.tractusx.bpdm.pool.mapper.shared.outbound.DataSpaceParticipantParseErrorMapper
 import org.eclipse.tractusx.bpdm.pool.model.parseAndExecuteAllOrNone
-import org.eclipse.tractusx.bpdm.pool.service.operation.DataSpaceParticipantUpdateService
-import org.eclipse.tractusx.bpdm.pool.service.parser.DataSpaceParticipantUpdateParser
+import org.eclipse.tractusx.bpdm.pool.service.operation.participation.DataSpaceParticipantUpdateService
+import org.eclipse.tractusx.bpdm.pool.service.parser.participation.DataSpaceParticipantUpdateParser
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.eclipse.tractusx.bpdm.pool.api.model.request.DataSpaceParticipantUpdateRequest as DataSpaceParticipantUpdateRequestDto