+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/service/application/v7/LegalEntitySearchApplicationV7Service.kt
@@ -26,8 +26,8 @@ import org.eclipse.tractusx.bpdm.common.service.toPageRequest
 import org.eclipse.tractusx.bpdm.pool.api.model.response.LegalEntityWithLegalAddressVerboseDto
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv7.inbound.LegalEntitySearchRequestMapper
 import org.eclipse.tractusx.bpdm.pool.mapper.poolv7.outbound.LegalEntityResponseMapper
-import org.eclipse.tractusx.bpdm.pool.service.operation.LegalEntitySearchService
-import org.eclipse.tractusx.bpdm.pool.service.parser.LegalEntitySearchParser
+import org.eclipse.tractusx.bpdm.pool.service.operation.legalentity.LegalEntitySearchService
+import org.eclipse.tractusx.bpdm.pool.service.parser.legalentity.LegalEntitySearchParser
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import org.eclipse.tractusx.bpdm.pool.api.model.request.LegalEntitySearchRequest as LegalEntitySearchRequestDto